feat: add Core ECDH-ES crypto logic
This is a second part of the ECDH-ES KW + AEAD crypto change.
It includes the core crypto logic for key wrapping and AEAD encryption
In a subtle package. This package is not meant to be used outside of ecdhes
package. Its use will be relavant in the next change.

Key wrapping is considered complete as part of this change.

part of ##1469
closes  #1472
closes  #1470

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>