fix: #2078 #2112
Set public key id to base58(pubSigKey) in vdri.Registry.Create() when
using legacy KMS. This sets the key's ID in the resulting did.Doc in
the proper format. This enables compatibility with the DID public key
fetcher from the `verifiable` package (format must be "<did#keyID>").

Strip a leading "#" (if any) when searching for public sig keys in the
BaseKMS' keystore to account for the above change.

Both these changes are done in code paths where legacy KMS is used and
will (presumably) be eliminated once we deprecate it.

Signed-off-by: George Aristy <george.aristy@securekey.com>