refactor: revert KID encoding back to base64URL
This change brings back base64URL key ID encoding for the user of
the store.

It introduces a new store wrapper used by the the KMS store and
the thridPartyKey store in Authcrypt/Packer that stores key IDs
base58 encoded and returns base64URL encoded IDs to the user of
the store.

closes #2137

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>