fix: Fixes mem store
The current implementation of mem storage has a bug in the Iterator method.
This bug can be reproduced only when storage is populated with the data.
The issue comes out when we are iterating over keys and the start key doesn't match.
In that case, the start index is equal to 0. Which means that the iterator will return data starting from 0 index. e.g data[0:end_idx_match].

This PR fixes it by introducing zero index as a -1.
PR includes:

Bugfix
Common test case (for all DBs)
Test case for mem implementation.

Signed-off-by: Andrii Soluk <isoluchok@gmail.com>