fix: unit tests depending on docker
Unit tests for storages were all ignoring failures (such as CouchDB or
MySQL not running).

This patch only ignores these tests when running in GitHub actions using
the 'macOS-10.15' virtual environment. Issue #2183 was opened to track
the issue of not running these tests in the macOS CI in the first place.

Signed-off-by: George Aristy <george.aristy@securekey.com>