refactor: add Wrap/Unwrap crypto functions
This change introduces Wrap() and Unwrap() to Cryto api.

The functions work for ECDHES keys only for now. Other keys
(ECDH1PU, Ed25519) can be introduced in a future change.

It also includes setting APU and APV in the key derivation
process (they are set to empty `[]byte{}` in the current JWE
building service).

closes #2257

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>