feat: add HeadersOpt optional function to remoteKMS New()
This change introduced a new optional function passed to remotekms.New()
This new option allows calling a function to potentially call an
authorization server and setup auth info as http headers. This function
should be created by the user of the framework. The absence of this option
means the remoteKMS is not setting any authorization info in the headers.

This change also removed 'secret' and 'controller' http headers as these
should not be dictated by the framework.

It also includes some minor refactoring like placing CreateKeystore()
above New() as a keystore is required before creating a new remoteKMS instance.

part of #2307

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>