feat: Eliminate unnecessary query calls in EDV REST provider
In an earlier commit, the EDV document IDs were changed so that instead of being randomly generated, they're instead based off hashes (computed by an encrypted hash function - HMAC) of the friendly keys. This means that querying the vault to get the document IDs in the EDV REST provider Get and Delete methods is no longer required, which will provide a big performance improvement.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>