fix: single recipient JWE with AAD decrypt error
This fixes a problem when using Flattened serialization of a JWE (using a single recipient).
Decryption fails when AAD is not empty because the authentication data is not properly
rebuilt when calling Decrypt().

This change also increases unit-test cases to cover Full (with AAD), Flattened (with AAD) and Compact(no AAD) serializations.

Closes #2496

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>