fix: create local doc cache loader
This change creates a local JSON-LD Document Cache loader using a cache of type *sync.Map
to avoid 'error on cache writes' panics.

This change also introduces the 'benchmark' Make target to support benchmarking code in the framework.

closes #1833 for the fatal error mentioned in the comment
Also related and closes the discussion about the same panic topic in #2487

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>