fix: ineffective didexchange.Client.CreateConnection()
PR #2588 changed the way `myDID` and `theirDID` were fetched before invoking
the message handlers: they are now fetched from the `did.ConnectionStore`.
This created misalignment with `didexchange.Client.CreateConnection()`
since it was not saving myDID nor theirDID to the `did.ConnectionStore`,
rendering it ineffective and resulting in an error once an agent
attempted to send a DIDComm message from a connection created using this
method.

Signed-off-by: George Aristy <george.aristy@securekey.com>