feat: support packer selection based on destination MediaTypes
This change sets the media type priority based on the destination according to Aries RFCs 19, 44 and 587.
It sets the JWE message's CTY header based on the received media type.
It also updates the default authcrypt packing sender key type to be AES256CBCHS512.
Finally, it resets optional connection logic when packing to allow OOB messages to be sent without a prior connection.

closes #1112 (packer is selected based on mediaTypeProfile)
closes #272 (key type is managed by the kms)
closes #749 (diffrent keys are supported by the packers, except LegacyPacker that supports ED25519 keys only)
closes #2858 (SendToDID now can be used without a DID connection, as it was before)

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>