feat: support DID Doc KeyAgreement.ID as kid/skid for DIDCommV2
This change adds DIDCommV2 support to use KeyAgreement.ID when a DID doc is available.
Default behaviour is to use did:key for pre DIDCommV2 or where DID doc is not available to the recipient.
The Peer VDR creator now uses DIDCommV2 service type matching the media porfile version: DIDCommMessaging (for V2).
Default is did-communication (for pre V2) if media profile is missing or is not DIDCommV2.

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>