docs: Updated storage interface documentation
Updated the documentation for the Put method to indicate that in the set of tags used on a single Put call, those tags must have unique tag names. The reason for this is because several of our major storage implementations in aries-framework-go-ext have this implementation, so this note was added to help ensure consistent behaviour.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>