feat: Full EDV provider support for AND+OR multiple tag queries
The EDV storage provider can now do full AND+OR tag querying. The SetStoreConfig method no longer calls the "add index" endpoint as it won't be added to the spec (indexing will be automatically by the EDV server).

Also updated storage interface documentation for the Query method to accommodate the new functionality.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>