fix: Docker warning when using frapsoft/openssl on arm64 system
Resolved a warning from Docker that would get printed when running the generate-test-keys Makefile target on an arm64 system. The warning from Docker alerts you that the image for frapsoft/openssl is for amd64, which doesn't match the system you're on (when using an arm64-based OS). To resolve the warning, you have to either use an image that matches the system architecture, or explicitly state the platform using the --platform flag. In this case, there is only an amd64 version of frapsoft/openssl, so I added the explicit flag to resolve the warning. I also added a TODO for us to find an arm64 alternative in the future (although the amd64 version of frapsoft/openssl does work fine on arm64 macOS currently as it seems that Docker makes use of QEMU to support emulation of non-native architectures).

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>