feat: support secp256k1 curve in KMS
This change adds secp256k1 keys support in the kms for the IEEE-P1363 format only.
DER format is not supported because the x509 package does not support this curve.
The framework does support the creation of an secp256k1 key with DER format, however storing it in the KMS is not supported.

closes #3410

Signed-off-by: Baha Shaaban <baha.shaaban@securekey.com>