refactor: Minimize interfaces required by didsignjwt.SignJWT
Minimized the interfaces required by the didsignjwt.SignJWT method by replacing the usage of the kms.KeyManager, crypto.Crypto, and vdr.Registry interfaces with minimized interfaces that define only the functionality needed for the SignJWT function. This makes it easier for consumers to inject in their own implementations. The method signatures of the new minimized interfaces still line up with the method signatures of the kms.KeyManager, crypto.Crypto, and vdr.Registry interfaces, so implementations of those interfaces can still be used without making any changes.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>