fix: didsignjwt Signer interface not exported
Fixed an issue with the didsignjwt.SignerGetter function signature that's used as a parameter to the SignJWT function. Since the return type of SignerGetter is an unexported interface, it means that external packages could not create a custom SignerGetter function and pass it in (which was the intent). This PR simply exports the Signer interface to make this possible.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>